% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pressure.R
\name{plot.pressure}
\alias{plot.pressure}
\alias{plot.pressure.casal2MPD}
\alias{plot.pressure.casal2TAB}
\title{plot.pressure plot fishing pressure if there has been an exploitation process reported.}
\usage{
plot.pressure(model, report_label = "", fisheryLabels = NULL, plot.it = T, ...)

\method{plot.pressure}{casal2MPD}(model, report_label = "", fisheryLabels = NULL, plot.it = T, ...)

\method{plot.pressure}{casal2TAB}(model, report_label = "", fisheryLabels = NULL, plot.it = T, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions using the Casal2 base library}

\item{report_label}{<string>}

\item{fisheryLabels}{if you only want to plot a subset of the fisheries supply a vecor of characters which corresponds to the fisheries you want to plot.}

\item{plot.it}{Whether to generate a default plot or return the values as a dataframe for personal plots}

\item{...}{remaining plotting options}
}
\value{
generate a plot over time if plot.it = T, if plot.it = F it will return a matrix of values.

\code{NULL}

\code{NULL}
}
\description{
A plotting function to plot fishing presuure (U's ) for 'casal2TAB' and 'casal2MPD' objects.
}
\details{
If you have multiple time-steps and fisheries happening at different time-steps it may be useful to use the fisheryLabels command to split out the plots.
}
\author{
Craig Marsh
}
